pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.82.0",
    "markdown>=3.10.2",
    "matplotlib>=3.10.8",
    "onnxruntime>=1.24.2",
    "optimum[onnxruntime]>=2.1.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "weaviate-client>=4.20.3",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
exclude = ["weaviate-docs"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
